Refactor into modular source files
